data-service/requirements.txt
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
connexion==2.14.1
Flask==2.2.2
google-auth==2.27.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==3.11.2
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
swagger-ui-bundle==0.0.9
urllib3==2.2.0
websocket-client==1.7.0
Werkzeug==2.3.8
